requirements.txt
absl-py==0.7.1
altgraph==0.16.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
autopep8==1.4.3
backcall==0.1.0
bleach==3.1.0
cffi==1.11.5
chardet==2.3.0
cluster==1.4.1.post2
colorama==0.4.1
common==0.1.2
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.2
django-bootstrap3==11.0.0
django-cors-headers==2.5.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fuel==0.2.0
future==0.17.1
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lightgbm==2.2.2
Markdown==3.1
MarkupSafe==1.1.0
matlab==0.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
mpmath==1.1.0
music21==5.5.0
mysql-connector==2.1.6
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numexpr==2.7.0
numpy==1.16.3
opencv-contrib-python==3.4.2.17
openpyxl==2.6.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pbr==5.1.3
pefile==2019.4.18
picklable-itertools==0.1.1
pickleshare==0.7.5
Pillow==5.3.0
progressbar2==3.47.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.7.1
pycodestyle==2.4.0
pycparser==2.19
pydot==1.4.1
pygame==1.9.6
Pygments==2.3.1
PyInstaller==3.5
pylint==2.2.2
PyMySQL==0.9.3
pyparsing==2.3.0
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.9.1
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.10.0
sklearn==0.0
sqlparse==0.3.0
ssh-import-id==5.5
statsmodels==0.10.1
sympy==1.4
tables==3.6.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==0.6.0
torch==1.3.0+cu92
torchvision==0.4.1+cu92
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.1
urllib3==1.13.1
vtk==8.1.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.10.11
xgboost==0.82
xlrd==1.2.0
xlwt==1.3.0
